CHANGES: snmpd: PATCH: 2082726: correctly report NIC speeds higher than 4Gbit/s
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@17263 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/include/net-snmp/data_access/interface.h
@@ -176,7 +176,7 @@ typedef struct netsnmp_interface_entry_s {
 typedef struct _conf_if_list {
     const char     *name;
     int             type;
-    u_long          speed;
+    unsigned long long speed;
     struct _conf_if_list *next;
 } netsnmp_conf_if_list;
